Refactor accommodation add
