+++ src/houseRent/apps/authentication/urls.py
@@ -9,5 +9,4 @@ urlpatterns = [
     path('logout/', auth_views.LogoutView.as_view(), name='logout'),
     path('register/', views.register_user, name='register'),
     path('login/privatePolicy', views.private_policy, name='privatePolicy'),
-    path('accommodation/add', views.register_acommodation, name='registerAccommodation') 
 ]
