+++ src/houseRent/apps/authentication/views.py
@@ -1,11 +1,11 @@
 # views.py
 from django.shortcuts import render, redirect
 from django.contrib.auth import authenticate, login
-from .forms import RegisterAccommodation, RegisterUser, RegisterAddress
+from .forms import RegisterUser, RegisterAddress
 from django.contrib.auth.views import LoginView
 
 from .forms import LoginForm, GuestLoginForm
-from apps.core.models import Accommodation, CustomUser, Address
+from apps.core.models import CustomUser, Address
 from django.contrib import messages
 
 
@@ -53,22 +53,4 @@ def register_user(request):
 def private_policy(request):
     return render(request, 'authentication/privatePolicy.html')
 
-def register_acommodation(request):
-    if request.method == 'POST':
-        accommodation = Accommodation()
-        address = Address()
-        formAccommodation = RegisterAccommodation(request.POST, instance=accommodation)
-        formAddress = RegisterAddress(request.POST, instance=address)
-        if formAccommodation.is_valid() and formAddress.is_valid():
-            address.save()
-            accommodation.address = address  
-            formAccommodation.instance.owner = CustomUser.objects.get(id=request.user.id)
-            # Guarda la instancia del formulario en la base de datos
-            formAccommodation.save()
-            messages.success(request, 'Alojamiento registrado correctamente')
-            return redirect('home')
-    else:
-        formAccommodation = RegisterAccommodation()
-        formAddress = RegisterAddress()
-   
-    return render(request, 'accommodation/add.html', {'formAccommodation': formAccommodation,'formAddress': formAddress})
+
